let test case "NextBiggerNumber When Input Is 12 Then return 21" PASS

+++ KataTest/KataTest.cs
@@ -9,7 +9,7 @@ namespace KataBiggerNumberTests
         [TestCase(11, -1, TestName = "NextBiggerNumber When Input Is 11 Then return -1")]
         [TestCase(21, -1, TestName = "NextBiggerNumber When Input Is 21 Then return -1")]
         [TestCase(12, 21, TestName = "NextBiggerNumber When Input Is 12 Then return 21")]
-        public void NextBiggerNumber_Test(int input, int expected)
+        public void NextBiggerNumber_Test(long input, long expected)
         {
             var result = Kata.NextBiggerNumber(input);
